README.md: update doc following API change for new() and set_program()
We just updated the behaviour of the `new()` and `set_program()` methods
to build eBPF VM and change their programs, so that they return a
`Result()` possibly containing an error. Programs trying to access the
newly created VMs must now `unwrap()` or do something equivalent. Update
the README.md file in that regard.