Fix mislabelled mul64_imm and mul64_reg (#573)
* Fix mislabelled mul64_imm and mul64_reg

MUL64_IMM and MUL64_REG opcodes were both mislabelled in the comments as div64 which also stated that they perform an operation equivalent `/=`. This commit accurately labels them both as mul64 and performing an operation equivalent to `*=`.

* Fix comments for neg64

neg64 was also mislabelled as requiring two operands `dst, src` when it only takes in `dst`. This is now fixed.